User Auth (#5)
* task: Complete User creation endpoint & fixed user model

User model was lacking a password field, since it was a minor change it
didn't warrant its own branch

* feat: Create Games model & Games end point

Added basic games model. Also added a default image to use for default cases.

* feat: Create Games model & Games end point

Added basic games model. Also added a default image to use for default cases.

* feat: Complete User update endpoint

The function will work whether update fields are undefined or not

*BUG*: if a number is passed to a "type: String" field, Mongoose will do
a type conversion and parse it as a string. For example:
    "username": 12345
This should fail the validator since its a number, but instead mongoose
will write:
    "username": "12345"

* feat: Complete User CRUD endpoints --DRAFT--

Missing features are highlighted on Trello. Existing features should be
more thoroughly tested in the future

* fix: Add bio field & set max lenght

* fix: Change get /users to answer an empty query with all users

* fix: minor

* chore: Create login endpoint & initialize session config

* feat: Complete user auth middleware

Co-authored-by: Pumpel <46710520+Pumpelgit@users.noreply.github.com>